Add http server
The files are now served by an http server that can be accessed at
localhost (or externally with additional configuration). The server used
is the http-server npm package.

This change was immediately needed as local files cannot be
programmatically accessed due to security restrictions. The obj importer
programmatically loads certain resources so this change was needed to
make the obj objects and related resources local. As such, the windmill
object is now local instead of being fetched from the
webgl2fundamentals.org website.