Add multi-object import support
It is now possible to import multiple objects into the scene, there is
no support yet for scaling, rotation and translation, so it isn't so
useful yet, but it's evolving quickly.

The process of importing multiple objects consists of calling
parseAndLoadObj with the path to the obj file, it will return an object
with fields zNear, zFar, cameraPosition, cameraTarget, objOffset and
parts, then passing parts along with u_world to the
renderObjectFromParts function inside the render loop. The other fields
can also be useful when rendering only one object, their names should be
pretty self-explanatory, if in doubt check the original code on
https://webgl2fundamentals.org/webgl/lessons/webgl-load-obj-w-mtl.html.