Add development live server
The live-server package can be used to run a live server. This can be
useful for development purposes, as it will removes the need to manually
refresh the page after each change to the code. The live server can be
started by running the script "live" with 'npm run live'.

The live-server package was not added to package.json as it appears to
have several vulnerabilities that may not be fixed in the near future
considering it is only intended for development purposes. It was not
added to package.json so that npm audit won't report any vulnerabilities
that would have be ignored, potentially causing vulnerabilities that are
an actual problem to be missed or mistakenly ignored.