Add translation, rotation and scale parameters
The renderObjectFromParts function now has translation, rotation and
scale parameters. The renderObjectFromParts function was also renamed
to simply renderObject.

The scene now has two animated objects demonstrating the translation,
rotation and scale: the windmill and the chair.

With this commit, the fundamental bases of the project are covered and
proper development can now start on separate branches.