Add desk model to demonstrate arbitrary obj import
A desk model downloaded from the wild
(https://www.cgtrader.com/free-3d-models/interior/interior-office/stylized-office-desk-2e013221-e453-4571-a843-cee84a676480)
was added to the assets and imported into the scene to demonstrate
arbitrary models can work with the importer.

The steps to make the model work weren't as straightforward as they
would be with a proper wavefront obj importer, but they weren't that
complicated either. To make things work the model was opened into
Blender, the materials were simplified, in the case of the desk only the
metallic texture was removed and the model was exported. The file had to
be manually edited to simplify the bump (normal) texture reference as
the importer doesn't support multiplying the bump map and Blender will
export the file with a multiplication factor, even if it's 1. The
textures also had to be converted to jpg from tga as it doesn't seem tga
is supported.