Add a lamp that can face a certain direction
A lamp asset was added that can face a certain direction. The lamp
receives a Y rotation value (the yaw) and a Z rotation value
(the head's pitch). With this values, the lamp can be made to face a
certain direction, so to look at a specific point in the scene or to a
specific object.