Add global axis debug object
A debug object that represents the global axis was added. It is much
larger than the "axis" object at 5 meters in each direction (10 meters
by 10 meters by 10 meters total) and has arrows indicating the positive
direction of each axis. The arrows also serve as a ruler: small arrows
have 10 cm spacing and large arrows have 1 m spacing. A commented-out
renderObject call to add this object was added so that it can be easily
added to the scene when needed.