Improve some debug primitives
The debugCube, debugSphere, and debugCircle objects had some issues. The
debugCube was centered on the bottom rather than the center. The
debugSphere has 0.01 off on every axis and the extremes where the
circles meet were not on the mathematical sphere surface, causing the
sphere object to have size 1.99 instead of 2.0. The debugCircle was
rotated by 360/64 degrees, the dimensions of the circle appeared to be
off by 0.01 since the extremes on the X and Y axis were not in the
mathematical circle, rotating it by half the span of the edges made the
X and Y extremes be on the mathematical circle fixing this issue.