Choose lamp targets with the Hungarian algorithm
Previously the lamps would be iterated and the closest target would be
chosen. This caused lamps close to objects to be given a target farther
away if the object closer to the lamp was already targeted by another
lamp that had its target selected before it. This meant the choice of
lamp targets didn't appear natural.

The Hungarian algorithm is used to find the optimal assignment of lamps
so that the total distance between lamps and their targets is the
smallest possible.

The complexity Hungarian algorithm is O(n^3), which isn't overly
problematic since there are not that many lamps and targets, but this
makes it even clearer that the generation code should not run in every
frame.