Remove unused code
Some code the belonged to the original tutorial code that was used as a
base was removed. A CSS import was also removed, the CSS controlled some
parts of the page that aren't in the project since the project only has
a WebGl canvas. This CSS put a black border on the WebGl canvas, so it
wasn't completely unused, but that black border isn't desired in this
project.