Make the procedural generator run only once
The procedural generator was running on every frame. This didn't cause
any apparent issues since the generator is seeded with a seed fixed per
execution. However, it's a waste of resources to run on it every frame
since it will always produce the same result. This is especially
important considering the lamps use the Hungarian algorithm, which has a
time complexity of O(n^3).

All the procedural generation code was moved to the
generateProceduralScene function, where it probably should have been
from the beginning. In the current state this function is only called
in the beginning of execution, however, when the UI for the procedural
generator is implemented, it will be called whenever the user changes
any of the parameters.

In the current architecture, however, to change any of the debug options
the generator is called again. This is not ideal, since the debug
options don't affect the generated scene, but isn't a problem worth
fixing since the generation cost of the scene is small for something
that doesn't run on every frame. This is happening because the generator
receives the debug options and return all the objects to be rendered,
including the debug objects. Changing this would require changing the
architecture to something that returns each of specific kinds of debug
objects separately, which would add complexity to no real benefit
considering the relatively small cost of the generator.