Compress textures to 1k JPG
Many textures were in the PNG format and had a resolution that was much
higher than needed. This caused the page to take several seconds to
load. Most textures were compressed to 1k JPG, including textures that
were already in JPG format but had a resolution higher than 1k. The only
exception is the debugPlane texture which was not changed as it is
rendered very large and needs a higher resolution.