Fix minor bugs
The seed input was setting the seed to be it's event, rather than the
input value, causing the seed to not line up with the input value. The
immediate effect of that is that inputting a seed would generate a
different result than if the seed had taken that value randomly.

The camera Z position is now 2, instead of -2, which was causing the
scene to be seen from the back.