Add a zNear divisor
The demos and the final scene can now specify a zNear divisor so that
the near plane can be moved closer to the camera. This was needed as the
final scene has a main object so large that the zNear plane was clipping
objects several centimeters away from the camera.