Load objects asynchronously
Objects were previously loaded one after the other, technically
asynchronously but awaiting the previous object to be loaded before
starting to load the next one.

The object loading is now done in parallel with Promise.all. The whole
object loading has been improved and now we also have callbacks
available to know when the loading is done.

A side effect of this that is the reason object loading was investigated
and rewritten to be parallel is that the last object to be loaded will
no longer appear without textures in the first frame. This issue
occurred because the last object loaded was not awaited and the
rendering started before its textures had been loaded. This highlighted
the problem that objects were loaded sequentially and not in parallel.

As a result of this change the loading time has been greatly reduced and
now only takes a couple of seconds instead of several. As such, the
loading screen, which has an ellipsis animation that starts slow due to
performance reasons beyond our control plays unacceptably slowly for the
entirety of the loading process. As a result, since we now have a
callback that gives us the total number of objects that have been loaded
we can now have a progress indicator in the loading screen. This is a
change that will have to be implemented for better user experience.