Improve the loading screen
The loading screen now has a loading bar to indicate progress.

The loading text no longer has an animated ellipsis, as it was too slow
and inconsistent in the beginning, causing a bad experience. This is
especially problematic now that the loading screen is shown for much
less time since the loading times have been greatly improved by loading
objects in parallel. Instead of an animated ellipsis, the loading text
will now pulse to indicate progress in ongoing even if the loading bar
isn't moving.

The loading screen will now also fade out when the loading is complete,
instead of simply disappearing.